+++ core-tests/multiple-pattern-test.hs
@@ -1,0 +1,14 @@
+import Test.Tasty
+import Test.Tasty.HUnit
+
+main :: IO ()
+main = defaultMain $ testGroup "all"
+  [ testGroup "red"
+    [ testCase "square" $ pure ()
+    , testCase "circle" $ pure ()
+    ]
+  , testGroup "green"
+    [ testCase "square" $ pure ()
+    , testCase "circle" $ pure ()
+    ]
+  ]